requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
anyio==4.7.0
apache-airflow==2.10.4
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.1
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.0
argcomplete==3.5.3
asgiref==3.8.1
attrs==24.3.0
babel==2.16.0
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.0
Deprecated==1.2.15
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.12.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
lazy-object-proxy==1.10.0
limits==3.14.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
numpy==2.2.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.22.3
setproctitle==1.3.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.3.0
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.17.0
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
